Skeleton plugin for powerd
The fu-plugin-upower.c was copied and renamed to fu-plugin-powerd.c to
fit the project it will be used for. This CL then rids the fu-plugin-powerd.c
file of unecessary functions and variables and deleting whitespace. This
leaves the file with just enough functions to be enabled on fwupd and
permit all updates.

The upower README.md file was re-created to exclude information specific to upower
and include information on powerd and the specific interface it will
use.
The meson.build file specific to upower was also re-created to be powerd-specific and
call fu-plugin-powerd.c. Also, the .conf install function was removed to be consistent with
an upstream change as it was deemed unecessary.

The meson.build general to all fwupd plugins was edited to include the
powerd subdirectory as well, so that the plugin will be enabled upon
compilation.

The .quirk file has been removed since updates to fwupd will create
a general daemon.quirk file for upower and powerd to share.

FIXUP: Adjusted copyright format and whitespace in fu-plugin-powerd.c